Pass documents directory URL to worklet (#9)
Previously the OS platform was passed and the path for the example app's
corestore was hard coded. The hard coded value did not work on an actual
iOS device returning an `EPERM` error. Instead the Documents directory
for the app is passed (provided by `expo-file-system`) and the Bare
worklet creates a `autopass-example` directory for the `Autopass`
corestore.